Delay function takes in an unsigned long
The delay function takes in an unsigned long so we should store
it as an unsigned long.

+++ stepper.h
@@ -17,7 +17,7 @@ typedef struct
 {
 	uint8_t step;
 	uint8_t pin0, pin1, pin2, pin3;
-	uint8_t delay;
+	unsigned long delay;
 } stepper_t;
 
 void stepper_init(stepper_t *stepper, uint8_t pin0, uint8_t pin1, uint8_t pin2, uint8_t pin3);
